# openapi.yaml
components:
  schemas:
    http.TokenIssuanceResponse:
      properties:
        accessToken:
          type: string
        accessTokenTTL:
          type: integer
        editorUrl:
          type: string
        wopiSrc:
          type: string
      required:
        - accessToken
        - accessTokenTTL
        - wopiSrc
        - editorUrl
      type: object
...
    http.tokenRequest:
      properties:
        documentId:
          type: string
      required:
        - documentId
      type: object
info:
  contact:
    email: info@alkem.io
    name: Alkemio Foundation
    url: https://alkem.io
  description: |
    WOPI protocol host service for Collabora Online integration into
    the Alkemio platform. Provides token issuance, document editing
    (CheckFileInfo, GetFile, PutFile), locking, and WOPI discovery.
  license:
    name: EUPL-1.2
    url: https://opensource.org/licenses/EUPL-1.2
  title: Alkemio WOPI Service
  version: 1.1.0
openapi: 3.1.1
paths:
...
  /wopi/token:
    post:
      operationId: TokenHandler.ServeHTTP
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/http.tokenRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/http.TokenIssuanceResponse'
          description: OK
        "400":
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
          description: Bad Request
        "401":
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
          description: Unauthorized
        "403":
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
          description: Forbidden
        "404":
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
          description: Not Found
        "405":
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
          description: Method Not Allowed
        "422":
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
          description: Unprocessable Entity
        "500":
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
          description: Internal Server Error
        "503":
          content:
            text/plain; charset=utf-8:
              schema:
                type: string
          description: Service Unavailable
      summary: ServeHTTP handles POST /wopi/token.